+++ RELEASE_NOTES.md
@@ -10,7 +10,7 @@ The demo app now uses a local package, which makes it a lot easier to develop th
 ### ✨ New features
 
 * `SystemNotificationContext` has a new completion-based dismiss function.
-* `SystemNotificationMessageConfiguration` has new `iconTextSpacing`, `minWidth` and `titleTextSpacing` properties.
+* `SystemNotificationMessageConfiguration` has new `iconTextSpacing` and `titleTextSpacing` properties.
 * `SystemNotificationUIKitOverlay` is a new view that simplifies adding a system notification to a UIKit view.
 
 ### 💡 Behavior changes
@@ -22,6 +22,10 @@ The demo app now uses a local package, which makes it a lot easier to develop th
 ### 🐛 Bug fixes
 
 * This version fixes a bug, where the message configuration padding was incorrectly applied.
+
+### 💥 Breaking changes
+
+* `SystemNotificationConfiguration` `minWidth` is no longer used and has been removed.